+++ frontend/src/metabase/query_builder/components/DataSelector/DataSelectorSchemaPicker/DataSelectorSchemaPicker.styled.tsx
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled";
+
+export const DataSelectorSchemaPickerContainer = styled.div`
+  width: 300px;
+`;
